CPU convert to state machine
Move the 6502 CPU to state machines.

Use 1 state machine to deal with reset and fetching of the 6502 ROM vectors.

Use 1 state machine to deal with the CPU execution. Avoid extra states when using immediate values/direct addresses.

IO_CPU add in a CLEAR queue value for UART and PS2, use in the beebeater MOS.